Extract `ExpressionUtils` to a file
